pyproject.toml
[project]
name = "aidlinkai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph-checkpoint-mongodb>=0.2.1",
    "pymongo>=4.15.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.11.0",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.33",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.0.0",
]
